1603_1.py (#35): --status-in-datapackage now is part of the drill (MVP implementation)

+++ officinam/999999999/999999999.lib.sh
@@ -2285,6 +2285,8 @@ temp_save_status() {
   _ex_librario="$ex_librario"
 
   status_archivum_codex="${ROOTDIR}/$_path/$_nomen.statum.yml"
+  datapackage_dictionaria="${ROOTDIR}/$_path/datapackage.json"
+  datapackage_librario="${ROOTDIR}/datapackage.json"
   status_archivum_librario="${ROOTDIR}/1603/1603.$_ex_librario.statum.yml"
   objectivum_archivum_temporarium="${ROOTDIR}/999999/0/1603+$_nomen.statum.yml"
 
@@ -2292,21 +2294,38 @@ temp_save_status() {
     --codex-de "$_nomen" --status-quo \
     >"$status_archivum_codex"
 
+  "${ROOTDIR}/999999999/0/1603_1.py" \
+    --codex-de "$_nomen" --status-quo --status-in-datapackage \
+    >"$datapackage_dictionaria"
+
   # echo "$status_archivum_librario status_archivum_librario "
 
-  set -x
+  # set -x
   "${ROOTDIR}/999999999/0/1603_1.py" \
     --codex-de "$_nomen" --status-quo --ex-librario="$_ex_librario" \
     >"$objectivum_archivum_temporarium"
-  set +x
+  # set +x
+
+  # We use statum.yml, not datapackage to know state of the library. That's why
+  # we can overryde directly
+  # set -x
+  "${ROOTDIR}/999999999/0/1603_1.py" \
+    --codex-de "$_nomen" --status-quo --ex-librario="$_ex_librario" \
+    --status-in-datapackage \
+    >"$datapackage_librario"
+  # set +x
 
   # NOTE: this operation should be atomic. But for sake of portability,
   #       we're using temporary file without flog or setlock or something.
   #       Trying to use --status-quo --ex-librario
   #       without temporary file will reset old information. That's why
   #       we're using temp file
-  rm "$status_archivum_librario" &&
+  if [ -f "$status_archivum_librario" ]; then
+    rm "$status_archivum_librario" &&
+      mv "$objectivum_archivum_temporarium" "$status_archivum_librario"
+  else
     mv "$objectivum_archivum_temporarium" "$status_archivum_librario"
+  fi
 
 }
